changed property type

+++ app/src/main/java/ch/hslu/mobpro/QuizIT/QuizITAPIService.kt
@@ -28,10 +28,11 @@ data class Question (
 )
 
 data class Score (
-    val timeInMilliseconds: Int,
+    val timeInMilliseconds: Long,
     val score: Int,
     val username: String
 )
+
 data class Leaderboard (
     val topTenScoresAsList: ArrayList<Score>
 )
